manifest-tree: add ability to move subtrees around for "diff"
Summary:
Add a way to register from->to subtree grafts that happen "temporarily" for any manifest diff operations. This can be used to perform diffs between different directories.

The general flow from Python will be:
  1. Some "directory branch" aware command will register the directory mappings on a context's manifest object at the start.
  2. Command proceeds, applying the grafts for the diff operation and diffing the directories.
  3. Command maps the paths produced by diff back to their original paths as needed (added in next commit).

This is a little weird, but in Python the manifest diff operation is often very deep in library code, so it isn't very practical to pipe CLI arguments to the manifest diff callsite, and then propagate the augmented diff info back up.

Reviewed By: quark-zju

Differential Revision: D60027039

fbshipit-source-id: 5a3804c9417db6cdefef0f69c340d134b81d1dff