manifest-tree: add "graft" operation
Summary:
Add operation to graft another node (either file or directory) into the current tree. This is different from other operations in that it intentionally overwrites any existing conflicts in the tree to make the graft.

This will be used by upcoming changes to effect "cross directory" operations by rewriting the manifest for diff operations.

Reviewed By: quark-zju, sggutier

Differential Revision: D60027042

fbshipit-source-id: 2020e0ced913f343876d1ce7b96d9b6bebdef1a8