derived_data_tailer: add sleep with exponential backoff when finding no new underived heads
Summary: Currently the derived data tailer repeatedly refreshes a bookmarks subscription which always makes a query to SQL master. This is fine for repositories that get a lot of traffic because we will be busy doing derivations, but for other repositories this means that we will repeatedly query SQL master and hammer xdb. This diff adds sleep with exponential backoff when refreshing the subscription results in no new underived heads.

Reviewed By: lmvasquezg

Differential Revision: D60116417

fbshipit-source-id: 47306b34059ee074ec6f588e6857ec1688c6e3c2