Add check fix for IncorrectWatchmanWatch
Summary:
# Context
Eden doctor is a tool used to try and fix eden when problems arise.

# This Diff
This diff implements the check_fix for IncorrectWatchmanWatch

# Technical Details
IncorrectWatchmanWatch is detected by checking that the watcher is "eden".
This check does the same.

This check is slightly redundant since the perform_fix also does the same check after trying the fix.

Reviewed By: genevievehelsel

Differential Revision: D59837115

fbshipit-source-id: fc9428e11968b1834f6d341cad3f0d6c29c2530b