merge: make manifestmerge() path-remapping aware
Summary:
To support operations such as grafting a commit between different directories, we need to be able to re-map directories so the diff operation is able to to diff different directories. For example `sl graft --from-path foo --to-path bar -r A` renames A's "foo" directory to "bar" so that it can be diffed with the "bar" directory of the working parent.

This means the diff operation produces paths like "bar/file", but we need to recover the original path "foo/file" to look up the file context in A, record the file path in the merge context, etc. To recover the original path, we use the `ungraftedpath()` manifest method.

I had to update the main loop in `manifestmerge()` to support different file paths for the "local", "remote" and "ancestor" contexts. I updated things with some time-travel insight of future diffs.

Reviewed By: quark-zju

Differential Revision: D60027033

fbshipit-source-id: 306d94886d538f8622c9a4c219056d4e1c7aed23