commit cloud: add remote bookmarks querying as map
Summary: Eventually in some commit cloud endpoints we need to constantly access the workspace's bookmarks  to check if a commit has any associated bookmarks, since we retrieve the bookmarks as a vec that would need to iterate over possibly all of it every single time. Instead let's add a helper to get this as a node -> bookmark map for easy and fast access.

Reviewed By: mitrandir77

Differential Revision: D59876003

fbshipit-source-id: cd75716800a32399fc56a9a0cfdcd5b5db4b3dd3