commit cloud: add local bookmarks querying as map
Summary: Eventually in some commit cloud endpoints we need to constantly access the workspace's bookmarks  to check if a commit has any associated bookmarks, since we retrieve the bookmarks as a vec that would need to iterate over possibly all of it every single time. Instead let's add a helper to get this as a node -> bookmark map for easy and fast access.

Reviewed By: RajivTS

Differential Revision: D59876004

fbshipit-source-id: 2f7d842a19f72eef7d71227e37c8afa0fa454d3e