Live Git Sync: test: Show that we derive exactly what we expect to derive
Summary:
Make the test more comprehensive so we catch any future regression.

* `fsnodes` is needed for the parent of any merge commit
* `skeleton_manifests` is needed for the parent of any stack when calling `repo_create_stack`
* `git_trees` is a dependency of `git_delta_manifests_v2`
* `git_delta_manifests_v2` is needed so we can serve the git protocol with the imported commits

All of these types must be derived in appropriately sized batches.

Nothing else should be derived at this stage.

To make this test work, we updated the jk value in the tests to derive `git_delta_manifests_v2`.
This is the prod setting we are converging towards (and already true for `chromium/src`)

Reviewed By: RajivTS

Differential Revision: D60121575

fbshipit-source-id: eb09ca24d198ffb9317cf3a8149acd592506cd2f