log sandcastle_alias in edenapi
Summary:
sandcastle_alias is very useful aggregaion in scuba because sandcastle jobs quite often can have the same set of identities, but can differ by alias a lot - workflows might be a lot different.

We would have solved sevs earlier if we had this.

Second of all this is a prerequisite for a T196176090

Reviewed By: singhsrb

Differential Revision: D60165049

fbshipit-source-id: 35060b12121238bfa0def7c2cfcd24497a75caf5