Standardize scuba logging for repo_stack_git_bundle_store endpoint
Summary: T196198076 was created because we thought we weren't logging the request and response parameters for the `repo_stack_git_bundle_store` endpoint. It turns out that we were already logging everything but it wasn't with the right prefix which is why we couldn't find this information at SEV time. This diff fixes that.

Differential Revision: D60166853

fbshipit-source-id: 73b0a94bc6390861b6cb0fe0aa575e102c1152ae