Provision for multiple transaction hooks during bookmark movement
Summary: The current API for bookmark transaction allows the caller to specify one hook into the transaction that can modify it before committing the transaction. However, we are trying to enable a multi-bookmark update operation which can have potentially multiple transaction hooks. This diff modifies the API so that it can handle multiple hooks into the main bookmark transaction

Reviewed By: andreacampi

Differential Revision: D60045269

fbshipit-source-id: 6462882569bc6ab1be553930fdffd56181be26b8