Extend Mononoke API to provide access to bookmark operations
Summary: In normal use cases, we don't really want access to the underlying bookmark operation and are just interested in the outcome of running the operation. However, when we want to link those operations together to make them atomic, we need to access the underlying operations. This diff achieves that.

Reviewed By: andreacampi

Differential Revision: D60054770

fbshipit-source-id: 1abcad9abdb70b3b376e95229c485e9622ca403b