Restructure Bookmark Movement API to support external transaction
Summary:
Currently, the `Bookmark Movement` API expects a single transaction to be committed as part of a single bookmark move since all the current Mononoke use cases follow that pattern. However, git has a `--atomic` setting that is [used in production](https://fburl.com/scuba/dev_command_timers/r3of5iei) which allows the user to update multiple refs atomically.

This diff is the first step in the stack to make that happen. It updates the API to accept an externally provided transaction that can be used to chain multiple operation together before committing them.

Reviewed By: andreacampi

Differential Revision: D60042303

fbshipit-source-id: 1041c765fa96999398a108844abfe92e43166b5e