diff: fix "a/" paths for xdir diffs
Summary: When diffing across directories, the "a/" path should match the "--from-path" directory, and the "b/" should match "--to-path". This was working for modified files, but not added/removed files. Fix by tracking existence of the file on each side with a seperate bool rather than setting that side's path.

Reviewed By: quark-zju

Differential Revision: D60054679

fbshipit-source-id: 270f182803864a72eb38e2a2c994b2c834136e8a