Fix stuck full paths mode
Summary:
Bug:
- on windows vscode
- hold ctrl key, see full paths of files
- continue holding, press tab to open quick file changer
- click on a different file or arrow key and release ctrl
- file is opened, but ISL didn't get the `keyup` event
- after navigating back to ISL, the full path view is stuck becuase we never got the keyup due to not having focus

I actually repro'd using opt-cmd-tab which is a slightly different picker but still exhibits the same behavior.

The fix I go for here is to reset all held keys on blur. This way, we're sure the paths will be reset when navigating back.

An alternate idea might be to reset the keys when ISL becomes hidden. But I think this same bug would happen from actions that unfocus ISL without hiding it entirely...

Reviewed By: quark-zju

Differential Revision: D60189671

fbshipit-source-id: 440f90b457a39b25ee9b61757bd6efea384c658c