edenapi: add retries for "health" and "capabilities"
Summary:
Requests can fail for lots of reasons, so we should always retry things that are safe to retry.

This was motivated by D60186397 where an error (Sapling's fault) on the first attempt to fetch the server capabilities was failing the entire "clone" operation. If we had retried it would of worked (but papered over the error, which would not have been ideal).

Reviewed By: quark-zju

Differential Revision: D60187014

fbshipit-source-id: 7f175533d18fc12294e0590061d7ae4ce023c462