shelve: exit unfinished merge state after interactive shelve
Summary:
According to the QA page [1], in some situations, sapling ends into the
confusing `The repository is in an unfinished *merge* state`. Jun introduced a
`sl continue` command before and we print a help message telling user to run
the command.

Intead of asking users to run `sl continue`, this diff just run the state
cleanup logic automatically at end of `shelve` command.

We also need to dive deep into the `goto` command to understand why it
results into this unfinished merge state. But for now, let's use this workaround to
improve user experience.

Reviewed By: muirdm

Differential Revision: D60157205

fbshipit-source-id: 61f35b7e6a6f47bf84f5f3718255bae25cf00ccd