clone: fix network doctor output on network error
Summary:
On clone, the config does not have remotefilelog.reponame set, so if there is a network error and cli dispatch invokes the network doctor, the network doctor was complaining about lack of remotefilelog.reponame instead of actually running.

Fix by allowing a default reponame. The debugnetworkdoctor command already set a default reponame to support running the network doctor outside a repo, so I just moved that into the network doctor library itself (and made it configurable).

Reviewed By: quark-zju

Differential Revision: D60188182

fbshipit-source-id: 748e4b8457a1424e0f8cab0be20f5facefc93424