hg_augmented_manifests: allow loading manifest with NULL hash
Summary: `HgAugmentedManifests` share the same hash as the corresponding hg manifest. So for empty hg manifests that have a NULL hash the corresponding `HgAugmentedManifests` will also have a NULL hash. Failing to load such `HgAugmentedManifests` means that we will start failing derivations for all of its descendants as shown in the integration test. To avoid this, this diff allows loading `HgAugmentedManifests` with NULL hash.

Reviewed By: gustavoavena

Differential Revision: D60170203

fbshipit-source-id: 110fcb1639db31240c510d0e2c1ef48ded4c6dde