globalrev_pushrebase_hook: clarify that the small repo is the globalrevs small repo
Summary: Clarify through better field name and comments that the small repo id in `GlobalrevConfig` is the id of the small repo that has globalrevs enabled.  Make it clear that we use the pushredirection config to determine where globalrevs are assigned.  If pushredirection is not enabled, they are assigned in the small repo.  If pushredirection *is* enabled, we assign them in the large repo, to all commits, whether they affect the globalrevs small repo or not.

Reviewed By: gustavoavena

Differential Revision: D60233916

fbshipit-source-id: 0333fecf414897e0e3bdd4a4d48390ae15d9e0e7