Fix handling of GHE without merge queue support (#926)
Summary:
Fix handling of GHE without merge queue support
https://github.com/facebook/sapling/issues/830 and https://github.com/facebook/sapling/issues/906 reported sapling not working for old Github Enterprise versions since their graphql schema doesnt include merge queue.

This adds a query to detect if merge queue is supported in the graphql version. Then, it will issue a different "YourPullRequestsQuery" depending on if merge queue is supported.

I considered a few other approaches:
- Just detect the error message and retry without merge queue field - a little gross to extract from the error
- Use "include" instead of an entirely separate query - didn't seem to count as a valid query still via graphiql at least (unsure if the query client here would strip out the unused field though)

Pull Request resolved: https://github.com/facebook/sapling/pull/926

Test Plan: Haven't had a chance to test yet, will try and validate that both versions work and that the detection logic works.

Reviewed By: evangrayk

Differential Revision: D60215245

fbshipit-source-id: cb36a77fbe0665353740d7fa97bc21ee9c86e0e8