Log exceptions/RemediationError from fixed problems in eden doctor
Summary:
This diff adds a new entry to the eden doctor log problem_failed_fixes_exceptions.
This field tracks any exceptions that were raised when running perform_fix or check_fix, including RemediationError

The intent of this is to be able to capture the details that are usually lost when a fix fails to succeed.

Since this is likely to consume alot of space, diff restricting that is here:
https://www.internalfb.com/diff/D60141827

Differential Revision: D60140939

fbshipit-source-id: a6ae99dcf907a62985e988f084807efdf7006f74