Add logging for eden doctor dry run
Summary:
This diff adds a new eden log type eden_doctor_dry_run.

The intent behind this new log is to track the success metrics of eden doctor improvments through the increased exposure of eden doctor results through the e-menu.

The log only fires on windows platforms for now to reduce the impact of the additional logs.

Differential Revision: D60180451

fbshipit-source-id: 60cb96cd582dea55b6d247b1315c6beb4a30fa33