Move public function that rewrites commits with submodule expansions
Summary:
Currently, we only support backsyncing and the public function from the `git_submodules` module that's responsible for syncing commits with expansion enabled is called `rewrite_commit_with_submodule_expansion` and is located in the `expand.rs` file.

I'm creating a new file, called `sync.rs`, moving it there and renaming it to `sync_commit_with_submodule_expansion`.

`sync_commit_with_submodule_expansion` will be responsible for determining when a sync is forwards or backwards and calling the appropriate functions in each case.

Reviewed By: RajivTS

Differential Revision: D60118877

fbshipit-source-id: 83d2f2394c8f557900817592625f012345090aa2