Use JK to guard execution of sharded x-repo-sync job
Summary: The x-repo-sync job has a constraint that only one instance of it should be running for a particular repo pair. Even though we can honor the constraint within multiple instances of the sharded job, there is nothing in place to prevent synchronization between sharded and non-sharded variants of this job. This JK does that. Once we get proof that the sharded variant works as expected, we will get rid of this JK.

Reviewed By: YousefSalama, clara-9

Differential Revision: D60242473

fbshipit-source-id: 7ae09f3033b3f3d16d933451af4e0d956559159a