Structure the x-repo-forward syncer as a sharded job
Summary: For any job to be sharded, its implementation must be defined in terms of `ShardedProcess` and `ShardedProcessExecutor`. This diff does the heavy-lifting part of that work. In the next diff, I will introduce sharding arguments for this job and invoke the sharded variant if those are provided. The current diff does not make a difference in the overall execution so existing integration tests should suffice.

Reviewed By: YousefSalama

Differential Revision: D60173042

fbshipit-source-id: cc36dff5b2dc0e0975e94f8f653f9f57d095a6b1