Enable leader-only execution of x-repo-sync job
Summary:
In the unsharded version of the job, we use `zkdaemon` in the TW spec and it works because every repo pair gets its own job repo region and `zkdaemon` helps ensure only one instance of the job runs. However, with the sharded version, there will be just one single job per region with multiple tasks contains multiple repo pairs. To ensure that we can still have the leader-only mode, this diff utilizes the `zk_leader_election` library.

NOTE: The current unsharded version of the job will not get impacted by this change. It will only affect the sharded job

Reviewed By: YousefSalama

Differential Revision: D60177789

fbshipit-source-id: f7c877fc59b7fed9d8dbb403ca102510dfde283e