Add AppModes to ISL for different root level rendering
Summary:
This stack will add a brand new webview panel in vscode that shows a dedicated comparison.

This was accessible previously as a modal when using ISL. The idea here is to make a separate webview, so you can look at comparisons without opening / focusing ISL. This also allows opening a comparison as a separate tab while using ISL, so you can keep it open on the side while you're doing something else.

To do this, we need to be able to render the ISL in different Modes. For example, "isl" mode will be regular ISL, "comparison" mode would be just a comparison viewer (not rendered as a Modal, but intended to be its own panel)

Reviewed By: quark-zju

Differential Revision: D59991635

fbshipit-source-id: 051a795985e80771ca9dd4ec304a81a3d3e5ca6d