Add shortcuts for comparison view panel
Summary:
Add same cmd-' and cmd-" shortcuts to open the panel form of the comparison view.

If ISL is focused, we still just use the modal comparison viewer

If ISL is not focused, the shortcut will use the vscode command to open the panel form

Note: This will conflict with the ISL1.0 comparison view shortcut. We'll need to disable that command after this one goes out.

Reviewed By: quark-zju

Differential Revision: D60188955

fbshipit-source-id: 80db9f86d4beb03120503d57487e29b691469e27