Add other comparison commands
Summary:
Similar to uncommitted comparison viewer, add "head" and "stack" as well.

Question: "head" may not really be the accurate term here. It's kind of inherited from ISL 1. In code we call the current commit `dot`, but I'm not sure what we've settled on as a user-facing name.

"last commit" could also be confusing. I just kept "head" for now. Open to thoughts here.

Reviewed By: quark-zju

Differential Revision: D60085282

fbshipit-source-id: 52d6f68f1b0b12544a1568f1358d6e65d9c06062