Add VSCodeServerPlatform
Summary:
Add a type for VSCodeServerPlatform, which contains the WebviewPanel or WebviewView for the ISL that this connection relates to.

Then, add the panel / view to this server  platform type, and pass it to createConnection.

This also involves updating `handleMessageFromClient` to know that the `this` context it'll be called with is going to be the ServerPlatform. This makes it easy to access / type without extra plumbing, since we already call it as `platform.handleMessageFromClient()` (thus setting the `this`)

Reviewed By: quark-zju

Differential Revision: D60085030

fbshipit-source-id: fd4d1d44753b8d875eb50c3056776e6f75a1f015