Add separate Comparison View vscode panel
Summary:
See previous diff for context.

This diff adds the vscode webview panel itself for the comparison. We expose a new vscode command that opens the comparison view. That command creates a new panel. This reuses some of the main ISL panel creation code, but makes sure to use a different ID. This means it does not have the same deserializing behavior (intentionally).

We also set the `window.islAppMode` in the HTML we generate. This is what allows the webivew <App> to change what is rendered at the root and only show the comparison viewer.

Note that our isl-server architecture supports this feature well thanks to its design. We can have multiple comparison views connect as normal ISL clients, with CWDs that can run commands and everything you'd expect.

----

There are more changes that will be needed here, including:
- remove the 'x' icon since it's not a modal that can be dismissed anymore
- add padding to the UI
- add the other comparison types as other commands
- add keyboard shortcuts to these commands
- phase out ISL1.0 equivalent commands in favor of this
- make the ISL UI know about these commands, and either use the Modal only when ISL is focused and otherwise use the separate panels, or have it be configurable with some kind of "pop out" UX, or something
- introduce comparison views for other embeddings of ISL, like `sl isl`?
- Add some kind of warning if the comparison you're looking at is obsolete. Now that it can be open for longer in the background, it's probably worth detecting when the set of files has changed and to warn you to reload the comparison.
- update our metrics & logging to know about different webview panels that are open. For example, ClientConnection should should include the AppMode

----

Performance note:
The way the code is currently set up, opening the comparison view will:
- create a webview panel like normal
- set the HTML like normal
- loads the entire isl webview.js file (somewhat large)
- notice its appMode is comparison, and render the <ComparisonViewApp>
- this suspends and uses React.lazy to load the actual <ComparisonView>. This is a side-effect of how the <ComparisonViewModal> is setup to do code splitting to reduce main bundle app size.
- the suspend finishes and the comparison view renders, and requests the server to do the comparions
- the comparison finishes and the result is sent down to the webview
- finally the actual diff view is shown

Notably, we could potentially improve performance:
- avoid the <Suspense> by reshuffling things so the comparison view is not loaded with React.lazy
- have a smaller initial load time by using code splitting to have the App entry point only include comparison view-related code and other essentials. This would reduce JS parsing time
- perhaps we could get the diff viewer to start running the sl diff command earlier and have the UI know to wait for the result.

The performance seemed adequate as a first approach, but I wanted to document the potential here.

Reviewed By: quark-zju

Differential Revision: D59991637

fbshipit-source-id: d16d9dc2c297491026c1841edfaae9e9b74dda52