Change panel title when changing comparison
Summary:
The new comparison viewer panel appears in its own panel with a title for the comparison, like "Uncommitted changes".

If you change the comparison mode in the dropdown, it should update the title. We can send a message to the platform to accomplish this, and now the server platform knows the panel / view and can access the `.title` to accomplish this.

Reviewed By: quark-zju

Differential Revision: D60085031

fbshipit-source-id: 3cd626f4cf067b2a016b3e77b1c6d1de103ccdd9