config: list sources for `%unset`s in built-in configs
Summary:
If we had an `%unset` in some of our built-in configs (e.g., `builtin_static/src/sapling.rs`), the source for this unset would not be listed.

We also have some things that depend on this behavior, such as the Rust config command. That command would sometimes it would output the wrong output because of this

Reviewed By: muirdm

Differential Revision: D60271114

fbshipit-source-id: ee05c54e805b26109bdae9b607900ca94eb21580