make RocksDbLocalStore async again
Summary:
# Context

We previously made the RocksDbLocalStore synchronous in D38551406 because it seemed to improve certain benchmarks (such as the glob benchmarks in the test plan of that diff). This is no longer the case on Linux, but it still holds true on macOS. However, the issue with making the RocksDbLocalStore synchronous is that on macOS we frequently wait on disk access in RocksDb. This causes us to block on reads very frequently, and can lead to NFS timeout popups when running certain eden prefetch commands.

This issue was observed during S433447 where users would run `arc focus` which would trigger a prefetch and then cause NFS popups to occur because Eden was not responding to NFS requests fast enough. Upon profiling, we noticed that the majority of time was spent blocked on RocksDb reads. We also discovered that turning off the LocalStore would mitigate the issue. However, through later benchmarking we discovered that the LocalStore was orders of magnitude faster than fetching from Sapling cache, and therefore we must keep it around for latency sensitive workflows.

This finally prompted us to make the RocksDbLocalStore asynchronous again to see how that impacted the workflows that were experiencing NFS popups. During testing, we noticed that Eden daemons using asynchronous RocksDbLocalStores would never encounter the NFS popups. Therefore, it seems like the positives from making the RocksDbLocalStore asynchronous vastly outweigh the negative performance impacts seen on the microbenchmark.

There's a chance that running RocksDbLocalStore actions async may flood the thrift executor with futures and cause an increase in queue timeout errors. There's an especially high risk for this with workflows such as checkoutRevision and getScmStatusV2 since these endpoints have high fanout.

In order to de-risk this change, we'll put the async change behind a config. We'll also add configs and the ability to run checkout/status serially to mitigate queue timeouts if we see an uptick in timeouts for endpoints with high fan out.

# This diff

Reads the EdenConfig to determine if RocksDbLocalStore actions should be done asynchronously or synchronously.

Reviewed By: genevievehelsel

Differential Revision: D60258737

fbshipit-source-id: fd3ec514a239af6a6232ad3ba40bfb4f74ba3a94