Add comparison panel VS Code setting
Summary:
This diff adds the vscode config setting for whether to always use a separate panel in vscode for the comparison view, or if it should use the modal if ISL is already focused.

Since this is a vscode-specific behavior, I think it's nice to make this into a vscode config setting. I guess this isn't totally necesary, and this could be in local storage. My thinking was that it's nice to have vscode-specific settings all collected in vscode settings, and not strewn between ISL setting tooltip and vscode config. Honestly, it would be nice UX to migrate our localstorage config settings to vscode configs. But that's probably out of the question for maintainability reasons. I digress.

The idea is to add a platform API to try to open a comparison in a dedicated window. Each platform can control this how they like, with settings or whatever. Here, we have the vscode platform use the vscode config to control this behavior. We cache the setting in the client via an atom, which we fetch on startup.

Note that in order to access this config atom in the vscode webview platform, we need to be careful about what gets imported when the platform is set up. So we need a dynamic import, even though this chunk will already be initialized in the bundle by the time we actually call this function. The platform must be initialized BEFORE everything else, including the server API / message bus. If the platform imports a file that uses serverAPI, then the message bus gets set up thinking the platform is "browser" which causes bugs.

Reviewed By: quark-zju

Differential Revision: D60303020

fbshipit-source-id: 7e655eb3baa90a63bf57d153648dbae52a4786dc