mercurial/mutation: add a test demonstrating abomonation alignment issues
Summary:
We use `Abomonation` for serialization of data to store in cache.  This library is very fast, but has some significant rough edges.  One such edge is that it doesn't provide any guarantees on alignment.  This means that if it entombs items that have higher alignment *after* items that have lower alignment, they may end up misaligned.  This can lead to panics, and potentially UB.

This diff adds a test which demonstrates the issue.  When run, it outputs something like:

```
thread 'caching::tests::abomonable' panicked at .../core/src/panicking.rs:220:5:
unsafe precondition(s) violated: slice::from_raw_parts_mut requires the pointer to be aligned and non-null, and the total size of the slice not to exceed `isize::MAX`
```

The test is left disabled for now until we fix the issue by forcing alignment.

Reviewed By: gustavoavena

Differential Revision: D60285064

fbshipit-source-id: cbe1d748d0c5b7f8e27ca876b1def387d9018fb6