Ensure the job can run in both sharded and unsharded mode
Summary:
The arguments for the x-repo-sync job include `SourceAndTargetRepoArgs` which are mandatory. However, when the job is running in sharded mode, these arguments will not be provided and instead the sharded service name and sharded service scope will be provided. This diff makes the necessary changes to ensure that the user can specify either set of arguments.

Ideally, this should have been handled through clap configuration but currently its incapable of doing without a bunch of refactor gymnastics.

Reviewed By: YousefSalama

Differential Revision: D60288184

fbshipit-source-id: 98c7ff10d79fb25c6f4e105f7b12a88f55ba124d