Use bookmark regex for x-repo-sync tailing from RepoConfig
Summary: Now that we have ported the `x-repo-sync` configuration from TW to configerator, let's update the code so that we read the bookmark regex from the config in the absence of command line argument. This will help maintain feature parity with non-sharded version of `x-repo-sync` job while allowing the sharded variant to use only the config. Once the non-sharded variant of the job is deprecated, we can decomission the CLI argument.

Reviewed By: YousefSalama

Differential Revision: D60376112

fbshipit-source-id: 329310c21ef18b69481288908e00358e5d4de73b