admin/derived_data: revert to using derive_exactly_batch in derive-slice subcommand
Summary: As part of killing `derived_data_utils`, I've recently changed the logic in the `monad derived-data derive-slice` subcommand to use `derive_bulk` instead of `derive_exactly_batch`. This is incorrect as `derive_bulk` assumes that if data is derived for one commit then it's derived for all its parents, which is not correct in this case as we derive boundaries beforehand. This diff goes back to using `derive_exactly_batch` to mitigate this.

Reviewed By: gustavoavena

Differential Revision: D60383176

fbshipit-source-id: c25c2f59c800578707737e3afa9d5c009e553e07