Fix creation if large_in_memory_repo in commit_syncer
Summary:
## This stack
Implementing support to backsync changes that touch the submodule expansions.

# This diff
Fixes something that isn't a bug now, but will be after backsyncing implementation.

We need to pass an `InMemoryRepo` version of the **large repo** to the `SubmoduleExpansionData` instance.
Now, we're passing the target repo because this code was written initially mostly assuming only forward sync.

We need to pass the actual large repo, regardless of the sync direction.

Reviewed By: RajivTS

Differential Revision: D60293842

fbshipit-source-id: 37a14dd71a8936205671cc51a4e9adfdbeaf2601