edenfsctl: add short option for --daemon-binary
Summary:
# Context

I use this option a lot when running development daemons. I could save a lot of time if there was a shorter option for specifying the daemon binary to run.

# This diff

This diff adds the short option and alters some tests to use the short option. No new tests were added since integration tests are expensive.

Reviewed By: kmancini

Differential Revision: D60307388

fbshipit-source-id: a31687a43858cf777aeaa6bde49d69d2bbf08975