website: change script thing location
Summary:
Currently Sapling has issues when running on shallow Git repos, even if the command being run is `debugshell`. This is causing our website [GitHub action to fail](https://github.com/facebook/sapling/actions/runs/10152255607/job/28073222395).

As a quick workaround for this, let's simply run the Sapling commands necessary to generate the website in the home directory instead of the directory of the script

Reviewed By: quark-zju

Differential Revision: D60418281

fbshipit-source-id: 145e1c7af7c8004bbbb63ee54e0d6e03e772f1e0