Implement getGlobFiles on FilteredBackingStore
Summary:
# Context
getGlobFiles is a BackingStore function that calls into the edenAPI function to retrieve a globFiles query from the Mononoke server. FilteredBackingStore is an implementation of BackingStore

# This Diff
Previously this was not implemented since filteredFS was still in development. Now that it has rolled out a little more, this function should be implemented.

# Technical Details
- Runs getGlobFiles in the backingStore
- Checks each path against the filter
- if the path is recursively filtered, drop it
- return a new result without the filtered paths

Reviewed By: MichaelCuevas

Differential Revision: D59981568

fbshipit-source-id: 987ed62b89b72e4360d3c3a225e1d504d310b750