Implement support for GitSha1 IDs in push error messages
Summary: The error messages in `git push` in Mononoke Git server are produced by the same `move_bookmark` code that is used in EdenAPI. As a result, the hashes in the error messages are `bonsai` instead of `git`. This diff adds the necessary changes to ensure that failures due to hooks and invalid user input gets reported with git hashes instead of bonsai.

Reviewed By: mitrandir77

Differential Revision: D60396622

fbshipit-source-id: 5ed688e69c5c66154c74fc034435207d23755a46