commit cloud: trait-based rewrite of the employee service
Summary:
~~So for integration tests we need to kinda mock this service so it works with disabled network access. This simply adds the endpoints used by `infer_workspace_identity` to the internal impl of the employee service (which is just a list of mocked employees) and wires up the configuration~~

New-ish diff: This diff is long but not very logically complex so I'll keep it for now, feel free to send any split requests.

The employee service so far was handled with two implementations: one for mock and for prod and each implementation lived side by side on each endpoint on different branches of a `match` statement. To fix this fragmentation, as gustavo suggested I re-wrote the whole thing as a trait-based polymorphism, maybe that way we could even plug in our own (or an oss) service aswell.

This only add the implementation, we will use this later in the stack

Reviewed By: gustavoavena

Differential Revision: D60202249

fbshipit-source-id: 535bcd0939efd1f4415e5269a51fefe9a83009e3