derived_data_manager: add derive_exactly_underived_batch method
Summary: The `derive_exactly_batch` method always derives the entire provided batch even if some of it is already derived. In some cases this is desirable to avoid extra blobstore/metadata DB fetches when the caller knows that the entire batch is underived or if the caller wants to rederive it, but in other cases (like `monad derived-data derive-slice` subcommand) this results in doing a lot of unnecessary work. This diff adds a `derive_exactly_underived_batch` method that first filter out derived changesets and then forwards to `derive_exactly_batch` to handle such cases.

Differential Revision: D60446363

fbshipit-source-id: dcdf84240712c13cc2d394c0adf442e488490a23