Use vscode setting for ligatures
Summary:
Ligatures are sometimes used in code fonts like fira code to show special glyphs for common coding symbols like `===` `=>`, `!=`, etc.

VS Code lets you control these ligatures to a decent degree with access to the `font-feature-settings` CSS property.

By default, vscode disables ligatures, but in webviews, HTML enables them by default.

So with special support we can always match vscode's behavior so the comparison view feels just like native vscode.

Note that we set this ligature value on the root `html` in css, insteadof only on editors. This is because we need to use `font-variant-ligatures: none` to disable ligatures, so using a custom CSS property doesn't seem to work well for us. I think this is ok but we can revist.

We also lightly validate this vscode config setting, since we set this directly into the webview html, though I don't think this is a particularly "real" attack vector.

Note also that we only set this at webview creation, so it takes a restart of ISL to take effect. I don't think this is worth the complexity for full live updating.

Reviewed By: sggutier

Differential Revision: D60424577

fbshipit-source-id: 37c080a2d4b4fc7c27a5436f9f280659181de272