Use vscode font size in comparison view
Summary:
Our comparison view was not correctly using the defined font size from vscode (if it exists). We should prefer that to match our comparison view to the rest of vscode.

Note: this only applies for the editor font size specifically. The global font size is actually controlled by the UI scale. I suppose there could be people who will dislike the editor font size being controll by the vscode setting but the UI scale affecting both the UI AND the editor font. We'll need to see if we need an extra way to opt out of that.

We also don't apply the line height because it seems in my testing that there isn't a line-height css property we can use from vscode. So this is just font size.

Reviewed By: sggutier

Differential Revision: D60419495

fbshipit-source-id: a8ed287315aa67fcd1aa0f77d46632eddda4fb4e